Fixed script to create 'dimensions.F' on Linux
It seems that 'read -d " "' behaves differently on macOS and Linux
(tested on Ubuntu 20.04 and Fedora 33) and on Linux will output empty
lines which don't appear on macOS. This change adds 'if' clauses to
ignore empty reads. The change has been tested on Linux and macOS and
works on both.