Updated CI workflow
This update will ensure that all branches and pull requests are tested,
regardless of branch. In addition it has simplified the Meson build step
by using a pre-configured Github action. Unfortunately this
pre-configured step can't be used with Intel since we need to ensure
that the Intel environment is sourced before build.