Update ci.yml with ecosys enabled (#284)
* Include ecosys in build command with intel cmpiler

* Restore restart_trcrd.F90 from release-1.3

Building BLOM stand-alone without iHAMOCC is does not allow any
references to hamocc functions. For now, it's more convenient to restore
'ifdef HAMOCC' when used in BLOM.

* Disable ecosys with gnu compilers, leave placeholder

Enabling ecosys for the gnu compiler build returns lots of strange error of the type:
```Error: Function ‘glb_inventory’ at (1) has no IMPLICIT type```
It seems to be a problem for the compiler when variables are imported from modules.