More transparent co2 gas-exchange formulation (#370)
This PR abandons the use of a fit for CO2 solubility that already includes corrections for non-ideality and water vapor (Weiss and Price 1980). Instead these corrections are now calculated explicitly (also following Weiss and Price 1980). This is slightly more accurate. However, the main reason for introducing this is that diagnostic quantities can be calculated in an easy and transparent way (e.g. the actual CO2 partial pressure at the air-sea interface, fugacity of CO2).